Move NAV & PCM code to Android LifecycleManager
